+++ src/main/java/com/potato/balbambalbam/main/cardList/dto/ResponseCardDto.java
@@ -1,15 +1,12 @@
 package com.potato.balbambalbam.main.cardList.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
+import lombok.*;
 
-@Setter
-@Getter
+@Setter @Getter
 @ToString
 @AllArgsConstructor
+@NoArgsConstructor
 @Schema(description = "카드리스트 각 카드 정보 Response")
 public class ResponseCardDto {
